requirements.txt
numpy==1.23.4
torch==1.11.0
torchvision==0.12.0
torchmetrics==0.7.2
opencv-python 
pandas>=1.3.5
Pillow==9.3.0
streamlit==1.9.0
matplotlib==3.5.1
einops
tqdm==4.63.0
scikit-learn==1.0.2
transformers==4.28.0